Make Linen and NNX logical rule deduction align
